% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kdtree.R
\name{find_np_kdtree}
\alias{find_np_kdtree}
\title{Find nearest point on the mesh for query coordinates using kdtree.}
\usage{
find_np_kdtree(query_coordinates, mesh, threads = parallel::detectCores())
}
\arguments{
\item{query_coordinates}{nx3 numerical matrix of x,y,z coordinates.}

\item{mesh}{an fs.surface instance, see \code{\link[freesurferformats]{read.fs.surface}} to get one.}

\item{threads}{integer, the number of threads to run in parallel.}
}
\value{
named list with keys 'coord' and 'distance'. 'coord': numeric nx3 kartesian coordinate matrix, the coordinates on the mesh which are closest to the \code{nx3} matrix of query_coordinates. 'distance': double vector, the distances to the respective coordinates in the 'coord' key.
}
\description{
Find nearest point on the mesh for query coordinates using kdtree.
}
\seealso{
\code{https://github.com/ThomasYeoLab/CBIG/blob/master/external_packages/SD/SDv1.5.1-svn593/BasicTools/MARS_findNV_kdTree.m}
}
\keyword{internal}
