% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kdtree.R
\name{find_nv_kdtree}
\alias{find_nv_kdtree}
\title{Find nearest mesh vertex for a coordinate using kdtree.}
\usage{
find_nv_kdtree(input_coordinates, mesh)
}
\arguments{
\item{input_coordinates}{nx3 numerical matrix of x,y,z coordinates.}

\item{mesh}{an fs.surface instance, see \code{\link[freesurferformats]{read.fs.surface}} to get one.}
}
\value{
integer vector, the \code{n} vertex indices which are closest to the \code{nx3} matrix of input_coordinates.
}
\description{
Find nearest mesh vertex for a coordinate using kdtree.
}
\seealso{
\code{https://github.com/ThomasYeoLab/CBIG/blob/master/external_packages/SD/SDv1.5.1-svn593/BasicTools/MARS_findNV_kdTree.m}
}
\keyword{internal}
