% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fwhm_smoothing.R
\name{get.avg.vertex.area}
\alias{get.avg.vertex.area}
\title{Compute the mean vertex area of a triangular mesh.}
\usage{
get.avg.vertex.area(surface = NULL, avg_vertex_area = NULL)
}
\arguments{
\item{surface}{a mesh, represented as an \code{fs.surface} instance from the \code{freesurferformats} package, a \code{tmesh3d} instance from \code{rgl}, or a character string representing the path of a surface file to load.}

\item{avg_vertex_area}{numeric scalar, the mean vertex area in squared mesh units (e.g., mm^2). If \code{NULL} (the default), it is computed from \code{surface}.}
}
\value{
numeric scalar, the mean (average) vertex area of the mesh in squared mesh units.
}
\description{
Compute the mean vertex area of a triangular mesh.
}
\keyword{internal}
