% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kdtree.R
\name{homogeneous_to_cartesian}
\alias{homogeneous_to_cartesian}
\title{Convert homogeneous coordinates to Cartesian coordinates.}
\usage{
homogeneous_to_cartesian(homog)
}
\arguments{
\item{homog}{nx4 numeric matrix of input coordinates}
}
\value{
nx3 matrix of Cartesian coordinates
}
\description{
Convert homogeneous coordinates to Cartesian coordinates.
}
\examples{
\dontrun{
homog = matrix(c(1,2,3,1,1,2,3,2), ncol=4, byrow=TRUE);
haze:::homogeneous_to_cartesian(homog);
}

}
\keyword{internal}
