% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kdtree.R
\name{homogenous_to_kartesian}
\alias{homogenous_to_kartesian}
\title{Convert homogenous coordinates to kartesian coordinates.}
\usage{
homogenous_to_kartesian(homog)
}
\arguments{
\item{homog}{nx4 numeric matrix of input coordinates}
}
\value{
nx3 matrix of kartesian coordinates
}
\description{
Convert homogenous coordinates to kartesian coordinates.
}
\keyword{internal}
