% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kdtree.R
\name{interp_tris}
\alias{interp_tris}
\title{Interpolate values within mesh triangles in R.}
\usage{
interp_tris(
  query_coordinates,
  mesh_vertices,
  nearest_face_vertices,
  pervertex_data,
  iwd_beta = 2
)
}
\arguments{
\item{query_coordinates}{nx3 numerical matrix of x,y,z coordinates. These are typically the vertex positions of a second (spherical!) mesh for that you need per-vertex data (e.g., the \code{fsaverage6} mesh).}

\item{mesh_vertices}{\code{nx3} matrix of x,y,z cartesian coordinates for the n mesh vertices (of the mesh that has the 'pervertex_data').}

\item{pervertex_data}{numerical vector, the continuous per-vertex data for the vertices of the mesh.}

\item{iwd_beta}{scalar double, the \code{beta} parameter for the inverse distance weight interpolation with the triangle. See details.}
}
\value{
vector with one entry per query coordinate (i.e., \code{nrow(query_coordinates)}).
}
\description{
Interpolate values within mesh triangles in R.
}
\keyword{internal}
