% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kdtree.R
\name{linear_interpolate_aux}
\alias{linear_interpolate_aux}
\title{Interpolate mesh per-vertex data at given points.}
\usage{
linear_interpolate_aux(
  query_coordinates,
  mesh_vertices,
  mesh_faces,
  vertex_neighbors,
  vertex_faces,
  query_coords_closest_vertex,
  pervertex_data
)
}
\arguments{
\item{query_coordinates}{nx3 numerical matrix of x,y,z coordinates.}

\item{mesh_vertices}{the vertex coordinates of the mesh that holds the pervertex_data}

\item{mesh_faces}{the faces (as indices into the vertex list) of the mesh that holds the pervertex_data}

\item{vertex_neighbors}{the adjacency list representation of the mesh (for each vertex: the vertices in distance 1 hop along mesh edges)}

\item{vertex_faces}{for each vertex, the list of face indices that the vertex is part of.}

\item{query_coords_closest_vertex}{the closest mesh vertex to the query coords, as computed with \code{find_nv_kdtree}}

\item{pervertex_data}{numerical vector, the continuous per-vertex data for the vertices of the mesh.}
}
\value{
named list with entries: 'interp_values', the numerical vector of interpolated data at the query_coordinates. 'nearest_vertex_in_face' the nearest vertex in the face that the respective query coordinate falls into, 'nearest_face' the index of the nearest face that the respective query coordinate falls into.
}
\description{
Interpolate mesh per-vertex data at given points.
}
\keyword{internal}
