% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kdtree.R
\name{linear_interpolate_kdtree}
\alias{linear_interpolate_kdtree}
\title{Map input values at points in 3D space onto mesh vertices.}
\usage{
linear_interpolate_kdtree(query_coordinates, mesh, pervertex_data)
}
\arguments{
\item{query_coordinates}{nx3 numerical matrix of x,y,z coordinates. These are typically the vertex positions of a second (spherical!) mesh for that you need per-vertex data (e.g., the \code{fsaverage6} mesh).}

\item{mesh}{a spherical fs.surface instance, see \code{\link[freesurferformats]{read.fs.surface}} to get one.}

\item{pervertex_data}{numerical vector, the continuous per-vertex data for the vertices of the mesh.}
}
\value{
named list with entries: 'interp_values', the numerical vector of interpolated data at the query_coordinates. 'nearest_vertex_in_face' the nearest vertex in the face that the respective query coordinate falls into, 'nearest_face' the index of the nearest face that the respective query coordinate falls into.
}
\description{
Map input values at points in 3D space onto mesh vertices.
}
\note{
The mesh must be spherical, and the query_coordinates must be located on the mesh sphere.
}
\seealso{
\code{https://github.com/ThomasYeoLab/CBIG/blob/master/external_packages/SD/SDv1.5.1-svn593/BasicTools/MARS_linearInterpolate_kdTree.m}

\code{https://github.com/ThomasYeoLab/CBIG/blob/master/external_packages/SD/SDv1.5.1-svn593/BasicTools/MARS_linearInterpolateAux.c}

\code{https://github.com/ThomasYeoLab/CBIG/blob/master/external_packages/SD/SDv1.5.1-svn593/BasicTools/MARS_findFace.m}
}
