% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_meshes_precomp.R
\name{mesh_neigh_pre}
\alias{mesh_neigh_pre}
\title{Return pre-computed neighborhood data for specific meshes.}
\usage{
mesh_neigh_pre(meshname)
}
\arguments{
\item{meshname}{a text identifier specifying the mesh you want connectivity data for. Currently supported meshes are listed here. 'lh_fsaverage': the left hemisphere of the FreeSurfer 6 fsaverage template. 'rh_fsaverage': the right hemisphere of the FreeSurfer 6 fsaverage template. 'lh_fsaverage6': the left hemisphere of the FreeSurfer 6 fsaverage6 template. 'rh_fsaverage6': the right hemisphere of the FreeSurfer 6 fsaverage6 template.}
}
\value{
list of vectors, the connectivity data as an adjacency list. The outer list has length n, where n is the number of vertices in the graph. The inner lists represent, for each vertex, all of its neighbors.
}
\description{
Return pre-computed neighborhood data for specific meshes.
}
