% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_gauss.R
\name{pervertexdata.smoothgauss}
\alias{pervertexdata.smoothgauss}
\title{Smooth per-vertex data based on a mesh using true Gaussian smoothing.}
\usage{
pervertexdata.smoothgauss(surface, pvdata, fwhm, trunc_factor = 3.5)
}
\arguments{
\item{surface}{a mesh, represented as an \code{fs.surface} instance from the \code{freesurferformats} package or a \code{tmesh3d} instance from \code{rgl}, or a character string representing the path of a mesh to load with \code{freesurferformats::read.fs.surface}.}

\item{pvdata}{numerical vector of per-vertex-data for the mesh, one value per vertex. Data values of \code{NA} will be ignored, allowing you to mask parts of the data. If you pass an \code{n x m} matrix or data.frame, the \code{n} rows will be treated as (independent) overlays that should be smoothed in parallel. To set the number of cores to use for parallel processing, set the 'mc_cores' option like this: \code{options("mc.cores"=22L);} before calling this function. Data.frames and matrices with a single row will be converted to vectors, and the return value will be a vector in that case.}

\item{fwhm}{scalar double, smoothing kernel full width at half max}

\item{trunc_factor}{scalar double, truncation factor for Gaussian neighborhood, in Gaussian standard deviations. Must be greater than 1.0 to make any sense, leave as is if in doubt. Larger values mean that a larger geodesic neighborhood needs to be computed on the mesh, which is slower.}
}
\value{
numerical vector, the smoothed data.
}
\description{
This function performs Gaussian smoothing of vertices in the geodesic neighborhood of all vertices. This is computationally quite expensive and requires large amounts of memory for large meshes.
}
\examples{
\dontrun{
mesh = rgl::tetrahedron3d();
pvd = rnorm(nrow(mes2$vb), mean = 5.0, sd = 1.0);
pvd_smoothed = pervertexdata.smoothgauss(mesh, pvd, fwhm=5.0);
}

}
\seealso{
\code{\link{pervertexdata.smoothnn}} can be used to approximate Gaussian smoothing with several iterations of nearest neighbor smoothing, and is a lot faster for large meshes.
}
