% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kdtree.R
\name{linear_interpolate_kdtree}
\alias{linear_interpolate_kdtree}
\title{Map input values at points in 3D space onto mesh vertices.}
\usage{
linear_interpolate_kdtree(query_coordinates, mesh, pervertex_data)
}
\arguments{
\item{query_coordinates}{nx3 numerical matrix of x,y,z coordinates.}

\item{mesh}{an fs.surface instance, see \code{\link[freesurferformats]{read.fs.surface}} to get one.}

\item{pervertex_data}{numerical vector, the continuous per-vertex data for the vertices of the mesh.}
}
\description{
Map input values at points in 3D space onto mesh vertices.
}
\seealso{
\code{https://github.com/ThomasYeoLab/CBIG/blob/master/external_packages/SD/SDv1.5.1-svn593/BasicTools/MARS_linearInterpolate_kdTree.m}

\code{https://github.com/ThomasYeoLab/CBIG/blob/master/external_packages/SD/SDv1.5.1-svn593/BasicTools/MARS_linearInterpolateAux.c}

\code{https://github.com/ThomasYeoLab/CBIG/blob/master/external_packages/SD/SDv1.5.1-svn593/BasicTools/MARS_findFace.m}
}
