% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_nn.R
\name{mesh.adj}
\alias{mesh.adj}
\title{Compute vertex neighborhoods for a mesh.}
\usage{
mesh.adj(surface, k = 1L)
}
\arguments{
\item{k}{scalar positive integer, the k value for the k-ring neighborhood. For k=1, this function computes the adjacency list representation of the graph (where the neighbors include the vertex itself).}
}
\description{
Compute vertex neighborhoods for a mesh.
}
