% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kdtree.R
\name{nn_interpolate_kdtree}
\alias{nn_interpolate_kdtree}
\title{Get per-vertex data at vertices closest to the given query coordinates on the mesh.}
\usage{
nn_interpolate_kdtree(query_coordinates, mesh, pervertex_data)
}
\arguments{
\item{query_coordinates}{nx3 numerical matrix of x,y,z coordinates. These are typically the vertex positions of a second (spherical!) mesh for that you need per-vertex data (e.g., the \code{fsaverage6} mesh).}

\item{mesh}{fs.surface instance, see \code{\link[freesurferformats]{read.fs.surface}} or \code{\link[fsbrain]{subject.surface}} to get one, or turn an \code{rgl} \code{tmesh} into one with \code{\link[fsbrain]{tmesh3d.to.fs.surface}}.}

\item{pervertex_data}{numerical vector, the continuous per-vertex data for the vertices of the mesh.}
}
\value{
the per-vertex data for the vertices closest to the query coordinates.
}
\description{
Return per-vertex data at the vertices closest to the given query points.
}
