% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_nn.R
\name{pervertexdata.smoothnn.adj.cpp}
\alias{pervertexdata.smoothnn.adj.cpp}
\title{Smooth data, C++ version.}
\usage{
pervertexdata.smoothnn.adj.cpp(mesh_adj, data, num_iter)
}
\description{
Smooth data, C++ version.
}
