% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth_nn.R
\name{pervertexdata.smoothnn.adj.cpp}
\alias{pervertexdata.smoothnn.adj.cpp}
\title{Smooth data, C++ version.}
\usage{
pervertexdata.smoothnn.adj.cpp(mesh_adj, pvdata, num_iter)
}
\arguments{
\item{mesh_adj}{list of vectors of integers, the adjacency list representation of the mesh. One can use the pre-computed adjacency for some special meshes, see \code{\link{mesh.neigh.pre}}. Data for vertices should include the vertex itself.}

\item{pvdata}{numerical vector of per-vertex-data for the mesh, one value per vertex. Data values of \code{NA} will be ignored, allowing you to mask parts of the data.}

\item{num_iter}{positive integer, number of smoothing iterations.}
}
\value{
numerical vector, the smoothed data.
}
\description{
Smooth data, C++ version.
}
\keyword{internal}
