% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_vv.R
\name{read.vv}
\alias{read.vv}
\title{Read vv binary file.}
\usage{
read.vv(filepath, datatype = integer())
}
\arguments{
\item{filepath}{string. Full path to the input vv file.}

\item{datatype}{on of `integer()` or `float()`.}
}
\value{
list of vectors, the data.
}
\description{
Read matrix-like data from vv files.
}
